refactor and update readme
