+++ README.md
@@ -33,41 +33,58 @@ pnpm add @zap/flash-langchain
 We welcome contributions! Please see our [Contributing Guidelines](CONTRIBUTING.md) for more information.
 
 ## Tools available 🧰
-> **34+ tools ready to use for your AI agents!**
+> **43+ tools ready to use for your AI agents!**
 
 ### Web3
 - [x] Ethereum
+  - [x] Transfer ETH
 - [x] Alchemy
-  - [x] Token Price APIs (by symbol, by address, historical)
-  - [x] Blockchain APIs (accounts, blocks, transactions)
-  - [x] Gas Estimation and Fee Management
-  - [x] Core Blockchain State (blocks, transactions, accounts)
-  - [x] Gas Management (estimateGas, gasPrice, feeHistory)
-  - [x] Smart Contract Interaction (logs, balances)
-- [ ] Coinbase Agentic Kit
+  - [x] Get Token Prices
+  - [x] Get Token Prices By Address
+  - [x] Get Historical Token Prices
+  - [x] Get Token Balances
+  - [x] Get Token Metadata
+  - [x] Get Token Allowance
+  - [x] Get ETH Balance
+  - [x] Get Transactions By Address
+  - [x] Get Block
+  - [x] Get Block Number
+  - [x] Get Transaction
+  - [x] Get Accounts
+  - [x] Get Chain ID
+  - [x] Get Logs
+  - [x] Get Gas Price
+  - [x] Estimate Gas
+  - [x] Get Fee History
 - [x] Hyperbolic Agentic Kit
-  - [x] GPU Management
-    - [x] Get Available GPUs
-    - [x] Rent Compute
-    - [x] Terminate Compute
-    - [x] Get GPU Status
-  - [x] Account Management
-    - [x] Get Current Balance
-    - [x] Get Spend History
-    - [x] Link Wallet Address
+  - [x] Get Available GPUs
+  - [x] Rent Compute
+  - [x] Terminate Compute
+  - [x] Get GPU Status
+  - [x] Remote Shell
+  - [x] SSH Access
+  - [x] Get Current Balance
+  - [x] Get Spend History
+  - [x] Link Wallet Address
+- [x] CoinGecko API Integration
+  - [x] Get Coin Prices
+  - [x] Get Coins List
+  - [x] Get Trending Coins
+  - [x] Get Coin Data
+  - [x] Get Coin Chart
+  - [x] Get Coin History
+  - [x] Get Coin Tickers
+  - [x] Get Coins Markets
+  - [x] Get Coins Market Data
+- [x] CoinGecko Pro API
+  - [x] Get Top Gainers & Losers
+- [ ] Coinbase Agentic Kit
 - [ ] Hedera Agentic Kit
 - [ ] Starknet Agentic Kit
-- [ ] P2P
-  - Staking
-  - Restaking
 - [ ] Solana 
-- [ ] Stripe
-  - Payment processing
+
 
 ### AI/ML
-- [x] Hyperbolic Agentic Kit
-  - [x] GPU Availability
-  - [x] Runtime Management
 - [x] ExaAI
   - [x] Web Search & Content Retrieval
 - [ ] ElevenLabs
@@ -75,14 +92,11 @@ We welcome contributions! Please see our [Contributing Guidelines](CONTRIBUTING.
 
 ### Productivity
 - [x] Google Integration
-  - [x] Gmail
-    - [x] List Gmail Messages
-  - [x] Calendar
-    - [x] Get Calendar List
-    - [x] Get Calendar Evencucuts
-    - [x] Insert Calendar Events
-    - [x] Delete Calendar Events
-  - [x] Sheets
+  - [x] List Gmail Messages
+  - [x] Get Calendar List
+  - [x] Get Calendar Events
+  - [x] Insert Calendar Events
+  - [x] Delete Calendar Events
 - [ ] Notion
 - [ ] Trello
 - [ ] Canva
@@ -108,10 +122,9 @@ We welcome contributions! Please see our [Contributing Guidelines](CONTRIBUTING.
 - [ ] Salesforce
 - [ ] HubSpot
 
-### Research
-- [ ] Youtube Scripts
-- [ ] Search Tool
-- [x] ExaAI
+### Research & Web
+- [x] ExaAI Search
+- [ ] YouTube Scripts 
 ## Implement Frameworks
 
 ### Javascript Frameworks
